bebob-protocols: maudio: special: move implementation of AV/C vendor dependent command for LED switch
